Refactor vocabulary-graph tests to self contained test suite
